follow ide advices

+++ javapng/src/main/java/com/sixlegs/image/png/CRCInputStream.java
@@ -21,7 +21,7 @@ extends FilterInputStream
 
     public long getValue()
     {
-        return (long)crc.getValue();
+        return crc.getValue();
     }
 
     public void reset()
